+++ packages/select/src/components/options-list/Component.tsx
@@ -114,14 +114,16 @@ export const OptionsList = forwardRef<HTMLDivElement, OptionsListProps>(
             );
         };
 
-        useVisibleOptions({
+        const actualOptionsCount = limitDynamicOptionGroupSize && options.length > 0;
+
+        const measured = useVisibleOptions({
             ...(!nativeScrollbar && { styleTargetRef: scrollbarRef }),
             visibleOptions,
             listRef,
             open,
             options,
-            actualOptionsCount: limitDynamicOptionGroupSize,
-            size: limitDynamicOptionGroupSize
+            actualOptionsCount,
+            size: actualOptionsCount
                 ? (() => {
                       switch (typeof size) {
                           case 'string':
@@ -165,6 +167,7 @@ export const OptionsList = forwardRef<HTMLDivElement, OptionsListProps>(
                     horizontalAutoStretch={optionsListWidth === 'content'}
                     scrollableNodeProps={scrollableNodeProps}
                     contentNodeProps={{ ref: listRef }}
+                    maskProps={{ className: measured ? undefined : styles.mask }}
                 >
                     {renderListItems()}
                 </Scrollbar>