+++ packages/status/src/docs/Component.stories.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import { select, text } from '@storybook/addon-knobs';
-import { Status, colors } from '@alfalab/core-components-status';
+import { boolean, select, text } from '@storybook/addon-knobs';
+import { Status } from '@alfalab/core-components-status';
+import { COLORS, SIZES } from '../consts';
 
 const meta: Meta<typeof Status> = {
     title: 'Components/Status',
@@ -15,8 +16,11 @@ export const status: Story = {
     name: 'Status',
     render: () => (
         <Status
-            color={select('color', colors, 'green')}
+            color={select('color', COLORS, 'green')}
             view={select('view', ['muted-alt', 'contrast', 'muted'], 'muted-alt')}
+            size={select('size', SIZES, 20)}
+            shape={select('shape', ['rectangular', 'rounded'], 'rectangular')}
+            uppercase={boolean('uppercase', true)}
         >
             {text('children', 'Label')}
         </Status>