+++ packages/tag/src/Component.responsive.tsx
@@ -1,6 +1,7 @@
 import React, { forwardRef } from 'react';
 
 import { useMatchMedia } from '@alfalab/core-components-mq';
+import { getComponentBreakpoint } from '@alfalab/core-components-shared';
 
 import { BaseTagProps } from './components/base-tag';
 import { TagDesktop } from './desktop';
@@ -20,7 +21,10 @@ export type TagProps = Omit<BaseTagProps, 'styles' | 'colorStylesMap'> & {
 };
 
 export const Tag = forwardRef<HTMLButtonElement, TagProps>(
-    ({ children, breakpoint = 1024, defaultMatchMediaValue, ...restProps }, ref) => {
+    (
+        { children, breakpoint = getComponentBreakpoint(), defaultMatchMediaValue, ...restProps },
+        ref,
+    ) => {
         const query = `(min-width: ${breakpoint}px)`;
 
         const [isDesktop] = useMatchMedia(query, defaultMatchMediaValue);