+++ packages/action-button/CHANGELOG.md
@@ -1,5 +1,40 @@
 # Change Log
 
+## 1.7.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.13.0
+
+## 1.7.1
+
+### Patch Changes
+
+<sup><time>13.09.2024</time></sup>
+
+### [#1370](https://github.com/core-ds/core-components/pull/1370)
+
+-   Заменили числовые значения на переменные отступов
+
+-   Обновлены зависимости
+    -   shared@0.12.1
+
+## 1.7.0
+
+### Minor Changes
+
+<sup><time>10.09.2024</time></sup>
+
+### [#1347](https://github.com/core-ds/core-components/pull/1347)
+
+-   Добавлена сборка moderncssm (ES2020, esm, сырые css-модули, отключен импорт базовых токенов)
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   spinner@3.5.0
+
 ## 1.6.8
 
 ### Patch Changes