+++ packages/alert/CHANGELOG.md
@@ -1,5 +1,48 @@
 # Change Log
 
+## 4.2.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   plate@7.7.0
+
+## 4.2.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   plate@7.6.3
+
+## 4.2.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   plate@7.6.2
+
+## 4.2.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   plate@7.6.1
+
+## 4.2.0
+
+### Minor Changes
+
+<sup><time>10.09.2024</time></sup>
+
+### [#1347](https://github.com/core-ds/core-components/pull/1347)
+
+-   Добавлена сборка moderncssm (ES2020, esm, сырые css-модули, отключен импорт базовых токенов)
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   plate@7.6.0
+
 ## 4.1.31
 
 ### Patch Changes