+++ packages/bank-card/CHANGELOG.md
@@ -1,5 +1,66 @@
 # Change Log
 
+## 5.8.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   masked-input@6.3.23
+
+## 5.8.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   masked-input@6.3.22
+
+## 5.8.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   masked-input@6.3.21
+
+## 5.8.1
+
+### Patch Changes
+
+<sup><time>13.09.2024</time></sup>
+
+### [#1358](https://github.com/core-ds/core-components/pull/1358)
+
+-   Обновлены наименования переменных скругления
+
+<sup><time>13.09.2024</time></sup>
+
+### [#1369](https://github.com/core-ds/core-components/pull/1369)
+
+-   Заменили числовые значения скругления на переменные
+
+<sup><time>13.09.2024</time></sup>
+
+### [#1370](https://github.com/core-ds/core-components/pull/1370)
+
+-   Заменили числовые значения на переменные отступов
+
+-   Обновлены зависимости
+    -   masked-input@6.3.20
+
+## 5.8.0
+
+### Minor Changes
+
+<sup><time>10.09.2024</time></sup>
+
+### [#1347](https://github.com/core-ds/core-components/pull/1347)
+
+-   Добавлена сборка moderncssm (ES2020, esm, сырые css-модули, отключен импорт базовых токенов)
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   masked-input@6.3.19
+
 ## 5.7.13
 
 ### Patch Changes