+++ packages/file-upload-item/CHANGELOG.md
@@ -1,5 +1,73 @@
 # Change Log
 
+## 5.10.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   icon-button@6.10.1
+
+## 5.10.0
+
+### Minor Changes
+
+<sup><time>20.09.2024</time></sup>
+
+### [#1386](https://github.com/core-ds/core-components/pull/1386)
+
+-   Откат обновления темизации corp из версии 47.16.0
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   icon-button@6.10.0
+
+## 5.9.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   icon-button@6.9.1
+
+## 5.9.0
+
+### Minor Changes
+
+<sup><time>13.09.2024</time></sup>
+
+### [#1360](https://github.com/core-ds/core-components/pull/1360)
+
+-   Обновление темы corp
+
+### Patch Changes
+
+<sup><time>13.09.2024</time></sup>
+
+### [#1370](https://github.com/core-ds/core-components/pull/1370)
+
+-   Заменили числовые значения на переменные отступов
+
+-   Обновлены зависимости
+    -   icon-button@6.9.0
+    -   link@5.3.1
+
+## 5.8.0
+
+### Minor Changes
+
+<sup><time>10.09.2024</time></sup>
+
+### [#1347](https://github.com/core-ds/core-components/pull/1347)
+
+-   Добавлена сборка moderncssm (ES2020, esm, сырые css-модули, отключен импорт базовых токенов)
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   icon-button@6.8.0
+    -   link@5.3.0
+    -   spinner@3.5.0
+
 ## 5.7.1
 
 ### Patch Changes