+++ packages/icon-view/CHANGELOG.md
@@ -1,5 +1,57 @@
 # Change Log
 
+## 3.8.3
+
+### Patch Changes
+
+<sup><time>11.10.2024</time></sup>
+
+### [#1399](https://github.com/core-ds/core-components/pull/1399)
+
+-   Изменена логика отрисовки image тэга - теперь image рендерится всегда и скрывается только в случае ошибки загрузки
+
+## 3.8.2
+
+### Patch Changes
+
+<sup><time>24.09.2024</time></sup>
+
+### [#1384](https://github.com/core-ds/core-components/pull/1384)
+
+-   Произвёл автоматическую оптимизацию кривых (по размеру) пакета 'icon-view'
+
+## 3.8.1
+
+### Patch Changes
+
+<sup><time>13.09.2024</time></sup>
+
+### [#1370](https://github.com/core-ds/core-components/pull/1370)
+
+-   Заменили числовые значения на переменные отступов
+
+## 3.8.0
+
+### Minor Changes
+
+<sup><time>10.09.2024</time></sup>
+
+### [#1347](https://github.com/core-ds/core-components/pull/1347)
+
+-   Добавлена сборка moderncssm (ES2020, esm, сырые css-модули, отключен импорт базовых токенов)
+
+### Patch Changes
+
+<sup><time>10.09.2024</time></sup>
+
+### [#1339](https://github.com/core-ds/core-components/pull/1339)
+
+Небольшой рефакторинг:
+
+-   оптимизация утилит (исправление копипасты и дубликаций)
+-   уточнение типов (исправление копипасты, вынос общего типа)
+-   уточнения в css (исправление shadow names)
+
 ## 3.7.1
 
 ### Patch Changes