+++ packages/markdown/CHANGELOG.md
@@ -1,5 +1,77 @@
 # @alfalab/core-components-markdown
 
+## 1.5.2
+
+### Patch Changes
+
+<sup><time>15.10.2024</time></sup>
+
+### [#1382](https://github.com/core-ds/core-components/pull/1382)
+
+-   Исправлено скругление картинок до `16px`
+
+## 1.5.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   typography@4.10.2
+    -   list@3.4.2
+
+## 1.5.0
+
+### Minor Changes
+
+<sup><time>20.09.2024</time></sup>
+
+### [#1213](https://github.com/core-ds/core-components/pull/1213)
+
+-   Добавлена возможность глобального переопределения breakpoint компонентов [Документация](https://core-ds.github.io/core-components/master/?path=/docs/instructions-breakpoints--docs)
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.13.0
+
+## 1.4.1
+
+### Patch Changes
+
+<sup><time>13.09.2024</time></sup>
+
+### [#1370](https://github.com/core-ds/core-components/pull/1370)
+
+-   Заменили числовые значения на переменные отступов
+
+<sup><time>13.09.2024</time></sup>
+
+### [#1358](https://github.com/core-ds/core-components/pull/1358)
+
+-   Обновлены наименования переменных скругления
+
+-   Обновлены зависимости
+    -   typography@4.10.1
+    -   link@5.3.1
+    -   list@3.4.1
+    -   shared@0.12.1
+
+## 1.4.0
+
+### Minor Changes
+
+<sup><time>10.09.2024</time></sup>
+
+### [#1347](https://github.com/core-ds/core-components/pull/1347)
+
+-   Добавлена сборка moderncssm (ES2020, esm, сырые css-модули, отключен импорт базовых токенов)
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   link@5.3.0
+    -   list@3.4.0
+    -   typography@4.10.0
+
 ## 1.3.2
 
 ### Patch Changes