+++ packages/navigation-bar-private/CHANGELOG.md
@@ -1,5 +1,96 @@
 # @alfalab/core-components-navigation-bar-private
 
+## 0.4.3
+
+### Patch Changes
+
+<sup><time>11.10.2024</time></sup>
+
+### [#1389](https://github.com/core-ds/core-components/pull/1389)
+
+-   Изменен тип у компонента navigation-bar-private для пропса title со string на ReactNode
+
+## 0.4.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   typography@4.10.2
+
+## 0.4.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.9.0
+    -   shared@0.13.0
+    -   icon-button@6.10.1
+
+## 0.4.0
+
+### Minor Changes
+
+<sup><time>20.09.2024</time></sup>
+
+### [#1386](https://github.com/core-ds/core-components/pull/1386)
+
+-   Откат обновления темизации corp из версии 47.16.0
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.8.0
+    -   icon-button@6.10.0
+
+## 0.3.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.7.1
+    -   icon-button@6.9.1
+
+## 0.3.0
+
+### Minor Changes
+
+<sup><time>13.09.2024</time></sup>
+
+### [#1360](https://github.com/core-ds/core-components/pull/1360)
+
+-   Обновление темы corp
+
+### Patch Changes
+
+<sup><time>13.09.2024</time></sup>
+
+### [#1370](https://github.com/core-ds/core-components/pull/1370)
+
+-   Заменили числовые значения на переменные отступов
+
+-   Обновлены зависимости
+    -   button@11.7.0
+    -   icon-button@6.9.0
+    -   typography@4.10.1
+    -   shared@0.12.1
+
+## 0.2.0
+
+### Minor Changes
+
+<sup><time>10.09.2024</time></sup>
+
+### [#1347](https://github.com/core-ds/core-components/pull/1347)
+
+-   Добавлена сборка moderncssm (ES2020, esm, сырые css-модули, отключен импорт базовых токенов)
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.6.0
+    -   icon-button@6.8.0
+    -   typography@4.10.0
+
 ## 0.1.10
 
 ### Patch Changes