+++ packages/navigation-bar/CHANGELOG.md
@@ -1,5 +1,65 @@
 # Change Log
 
+## 1.2.5
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   typography@4.10.2
+
+## 1.2.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.9.0
+    -   shared@0.13.0
+
+## 1.2.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.8.0
+
+## 1.2.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.7.1
+
+## 1.2.1
+
+### Patch Changes
+
+<sup><time>13.09.2024</time></sup>
+
+### [#1370](https://github.com/core-ds/core-components/pull/1370)
+
+-   Заменили числовые значения на переменные отступов
+
+-   Обновлены зависимости
+    -   button@11.7.0
+    -   typography@4.10.1
+    -   shared@0.12.1
+
+## 1.2.0
+
+### Minor Changes
+
+<sup><time>10.09.2024</time></sup>
+
+### [#1347](https://github.com/core-ds/core-components/pull/1347)
+
+-   Добавлена сборка moderncssm (ES2020, esm, сырые css-модули, отключен импорт базовых токенов)
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.6.0
+    -   typography@4.10.0
+
 ## 1.1.3
 
 ### Patch Changes