+++ packages/typography/CHANGELOG.md
@@ -1,5 +1,43 @@
 # Change Log
 
+## 4.10.2
+
+### Patch Changes
+
+<sup><time>04.10.2024</time></sup>
+
+### [#1401](https://github.com/core-ds/core-components/pull/1401)
+
+-   Добавлен `opacity` во время показа скелетона. Это позволяет избежать моргания компонента в ssr.
+
+## 4.10.1
+
+### Patch Changes
+
+<sup><time>13.09.2024</time></sup>
+
+### [#1370](https://github.com/core-ds/core-components/pull/1370)
+
+-   Заменили числовые значения на переменные отступов
+
+-   Обновлены зависимости
+    -   skeleton@5.2.0
+
+## 4.10.0
+
+### Minor Changes
+
+<sup><time>10.09.2024</time></sup>
+
+### [#1347](https://github.com/core-ds/core-components/pull/1347)
+
+-   Добавлена сборка moderncssm (ES2020, esm, сырые css-модули, отключен импорт базовых токенов)
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   skeleton@5.1.0
+
 ## 4.9.0
 
 ### Minor Changes